Move drawing functions to drawer object
